requirements.txt
ignite==1.1.0
lpips==0.1.4
matplotlib==3.6.2
monai==1.3.0
numpy==1.25.2
pandas==1.5.0
Pillow==10.4.0
scipy==1.14.0
torch==2.4.0+cu121
torchvision==0.19.0
tqdm==4.66.2
xformers==0.0.27.post2
